src/requirements_minimal.txt
# Core dependencies
python-dotenv==1.0.1
streamlit==1.42.0
pydantic>=2.0.0

# LangChain dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-text-splitters==0.0.1
langchain-anthropic==0.0.1

# NLP & AI Models
google-generativeai>=0.3.0
transformers==4.48.0
torch>=2.0.0
sentence-transformers==3.4.1

# Utilities
httpx==0.28.1
anthropic==0.8.1